+++ armyknife/on_aw.py
@@ -141,6 +141,8 @@ class OnAWSpark(object, on_aw.OnAWBase):
                     markdown=True)
                 spark.link.post_admin_message(
                     text="Successfully authorized account, but could not delete old firehose: " + email)
+                spark.link.post_admin_message(
+                    text=str(self.auth.oauth.last_response_code) + ':' + self.auth.oauth.last_response_message)
                 return True
         msghash = hashlib.sha256()
         msghash.update(spark.me.passphrase)
@@ -243,6 +245,7 @@ class OnAWSpark(object, on_aw.OnAWBase):
     def post_callbacks(self, name):
         if not self.myself or not self.myself.id:
             logging.debug("Got a firehose callback for an unknown user.")
+            self.webobj.response.set_status(410, 'Gone')
             return True
         spark = sparkrequest.SparkRequest(
             body=self.webobj.request.body,